+++ history.md
@@ -6,7 +6,7 @@
 - Little additions to both `README.md` files (the main one and the meicoPy).
 - Added method `meico.midi.Midi.writeMidi(String filename)`.
 - The `write...` methods in classes `meico.midi.Midi` and `meico.audio.Audio` did not ensure that the file actually exists in the file system or is created if not. This has been fixed.
-- All `write...` methods in classes `meico.midi.Midi` and `meico.audio.Audio` return a `boolean` on the success of the file writing. Exception throwing has been removed. Classes `meico.app.Main` and `meico.app.MeicoApp` have been adapted accordingly.
+- All `write...` methods in classes `meico.midi.Midi` and `meico.audio.Audio` return a `boolean` on the success of the file writing. Exception throwing has been removed. Classes `meico.app.Main`, `meico.app.MeicoApp` and `meicoPy.py` have been adapted accordingly.
 
 
 #### v0.2.18